requirements.txt
openai
python-docx
python-dotenv
azure-cognitiveservices-speech
